Add support for R4iTT's DSTT clones
PK3DS has not been tested, but since the R4 3DS Gold RTS is a clone of it - and that the PK3DS was made by R4iTT - I believe that it's safe to assume this will work on that cart too. These carts have issues using the ttio_sdhc DLDI, however they work just fine using Boyakkey's DSTT DLDI, hence why it's now in a seperate folder.